todos os erros e repostas estão padronizados

+++ app/routes/promocao_route.py
@@ -2,16 +2,26 @@ from flask import Blueprint, request, jsonify
 from app.services.promocao_service import PromocaoService
 from app.repositories.promocao_repository import PromocaoRepository
 from app.util.decorator_perfil import perfil_required
+from app.util.api_response import resposta_sucesso
+from app.util.api_error import ApiError
 
 promocao_bp = Blueprint('promocao', __name__)
 
 @promocao_bp.route('/admin/promocoes/criar', methods=['POST'])
 @perfil_required("GESTAO")
 def criar_promocao():
-    try:
-        dados = request.get_json()
-        promocao = PromocaoService.create_promocao(dados)
-        return jsonify({
+    dados = request.get_json()
+    if not isinstance(dados, dict):
+        raise ApiError(
+            error="JSON_INVALIDO",
+            message="O corpo da requisição deve ser um JSON válido.",
+            status_code=400,
+            details=[]
+        )
+    promocao = PromocaoService.create_promocao(dados)
+    return resposta_sucesso(
+        message="A promoção foi criada com sucesso.",
+        data={
             "id": promocao.id,
             "nome": promocao.nome,
             "produto_id": promocao.produto_id,
@@ -21,43 +31,70 @@ def criar_promocao():
             "quantidade_minima": promocao.quantidade_minima,
             "data_inicio": promocao.data_inicio.isoformat(),
             "data_fim": promocao.data_fim.isoformat(),
-            "ativa": promocao.ativa
-        }), 201
-    except ValueError as erro:
-        return jsonify({"erro": str(erro)}), 400
-    except Exception as e:
-        return jsonify({"erro": str(e)}), 500
-
+            "ativa": promocao.ativa            
+        },
+        status_code=201
+    ) 
 
 @promocao_bp.patch('/admin/promocoes/<int:promocao_id>/ativar')
 @perfil_required("GESTAO")
 def ativar_promocao(promocao_id):
-    try:
-        PromocaoService.atividade(promocao_id, True)
-        return jsonify ({
-            "mensagem":"promoção ativa"
-        }), 200
-    except Exception as e:
-        return jsonify({"erro":str(e)}), 400 
+    if promocao_id is None:
+        raise ApiError(
+            error="PROMOCAO_NAO_ENCONTRADA",
+            message="A promoção informada não foi encontrada.",
+            status_code=404,
+            details=[{
+                "field":"promocaoId",
+                "issue":f"A promoção com o Id {promocao_id} não existe."
+            }]
+        )
+    PromocaoService.atividade(promocao_id, True)
+    return resposta_sucesso(
+        message="A promoção foi ativada com sucesso.",
+        status_code=200
+    )
 
 @promocao_bp.patch('/admin/promocoes/<int:promocao_id>/desativar')
 @perfil_required("GESTAO")
 def desativar_promocao(promocao_id):
-    try:
-        PromocaoService.atividade(promocao_id, False)
-        return jsonify ({
-            "mensagem":"promoção desativada"
-        }), 200
-    except Exception as e:
-        return jsonify({"erro":str(e)}), 400     
+    if promocao_id is None:
+        raise ApiError(
+            error="PROMOCAO_NAO_ENCONTRADA",
+            message="A promoção informada não foi encontrada.",
+            status_code=404,
+            details=[{
+                "field":"promocaoId",
+                "issue":f"A promoção com o Id {promocao_id} não existe."
+            }]
+        )    
+    PromocaoService.atividade(promocao_id, False)
+    return resposta_sucesso(
+        message="A promoção foi desativada com sucesso.",
+        status_code=200
+    )
 
 @promocao_bp.get('/admin/promocoes')
 @perfil_required("ADMINISTRADOR", "GERENCIA", "GESTAO")
 def listar_promocoes():
-    try:
-        registros = PromocaoRepository.listar_ativas()
-        return jsonify([
-            {
+    filtros={
+        "promocao_id": request.args.get("promocaoId"),
+        "nome_promocao": request.args.get("nomePromocao"),
+        "produto_id": request.args.get("produtoId"),
+        "tipo_promocao": request.args.get("tipoPromocao"),
+        "data_inicio": request.args.get("dataInicio"),
+        "data_fim": request.args.get("dataFim"),
+        "valor_max": request.args.get("valorMax"),
+        "valor_min": request.args.get("valorMin"),
+        "ativa": request.args.get("ativa"),
+        "ordenacao": request.args.get("ordenacao", default="promocaoId_asc"),
+        "page": request.args.get("page", default=1, type=int),
+        "limit": request.args.get("limit", default=20, type=int)
+    }
+    registros = PromocaoService.listar_promocao_filtrada(filtros)
+    return resposta_sucesso(
+        message="A lista de promoções foi consultada com sucesso.",
+        data={{
                 "id": promocao.id,
                 "nome": promocao.nome,
                 "produto_id": promocao.produto_id,
@@ -69,20 +106,26 @@ def listar_promocoes():
                 "data_fim": promocao.data_fim.isoformat()            
             }
             for promocao in registros
-        ]), 200
-    except Exception as e:
-            return jsonify({"erro":str(e)}), 400 
-
+        },
+        meta=registros["meta"],
+        status_code=200
+    )
 
 @promocao_bp.delete('/admin/promocao/<int:promocao_id>/delete')
 @perfil_required("GESTAO")
 def delete_promocao(promocao_id):
-    try:
-        PromocaoService.delete_promocao(promocao_id)
-        return jsonify({
-            "mensagem": "Promoção excluído com sucesso."
-        }), 200
-    except ValueError as erro:
-        return jsonify({"erro": str(erro)}), 404
-    except Exception as erro:
-        return jsonify({"erro": str(erro)}), 500      
+    if promocao_id is None:
+        raise ApiError(
+            error="PROMOCAO_NAO_ENCONTRADA",
+            message="A promoção informada não foi encontrada.",
+            status_code=404,
+            details=[{
+                "field":"promocaoId",
+                "issue": f"A promoção com o Id {promocao_id} não existe."
+            }]
+        )
+    PromocaoService.delete_promocao(promocao_id)
+    return resposta_sucesso(
+        message="A promoção foi deletada com sucesso.",
+        status_code=200
+    )
